# services/manager/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SkausWatch Manager API
  description: SkausWatch manager service — users, alerts, threat intelligence, approvals, ENDPOINT agent fleet, SIEM, ASM, CodeScan proxy, S3 malware scanning, and research lookups.
  contact:
    name: Penguin Tech Inc
    email: support@penguintech.io
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: '1'
paths:
...
  /api/v1/alerts/search:
    post:
      tags:
      - alerts
      summary: |-
        POST /alerts/search — richer filtering than GET /alerts (free-text query,
        date range, `assigned_to`).
      operationId: search_alerts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchBody'
        required: true
      responses:
        '200':
          description: Paginated search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSearchResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/alerts/{alert_id}:
    get:
      tags:
      - alerts
      summary: GET /alerts/{alert_id} — single alert detail.
      operationId: get_alert
      parameters:
      - name: alert_id
        in: path
        description: Alert id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Alert detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertItem'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
    put:
      tags:
      - alerts
      summary: PUT /alerts/{alert_id} — admin/maintainer only; partial update.
      operationId: update_alert
      parameters:
      - name: alert_id
        in: path
        description: Alert id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBody'
        required: true
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertUpdateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/alerts/{alert_id}/status:
    put:
      tags:
      - alerts
      summary: |-
        PUT /alerts/{alert_id}/status — admin/maintainer only (contract defect
        #7: v1 let any authenticated role mutate status; this port tightens the
        gate to match the role model).
      operationId: update_alert_status
      parameters:
      - name: alert_id
        in: path
        description: Alert id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusBody'
        required: true
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertStatusResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: |-
        POST /auth/refresh — unauthenticated (the refresh token in the body is
        the credential; no bearer header is required or checked).
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Rotated access/refresh token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid, revoked, or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AlertItem:
      type: object
      description: Documentation-only mirror of `alert_json`'s wire shape.
      required:
      - id
      - title
      - severity
      - status
      - indicators
      properties:
        ai_review: {}
        assigned_to:
          type:
          - integer
          - 'null'
          format: int32
        created_at:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        id:
          type: integer
          format: int32
        indicators: {}
        resolution_notes:
          type:
          - string
          - 'null'
        resolved_at:
          type:
          - string
          - 'null'
        severity:
          type: string
        source:
          type:
          - string
          - 'null'
        status:
          type: string
        title:
          type: string
        updated_at:
          type:
          - string
          - 'null'
...
    AlertSearchItem:
      type: object
      description: Documentation-only mirror of `search_json`'s wire shape.
      required:
      - id
      - title
      - severity
      - status
      properties:
        created_at:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        id:
          type: integer
          format: int32
        severity:
          type: string
        source:
          type:
          - string
          - 'null'
        status:
          type: string
        title:
          type: string
    AlertSearchResponse:
      type: object
      description: Documentation-only mirror of `search_alerts`'s `serde_json::json!` body.
      required:
      - items
      - total
      - page
      - per_page
      - pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AlertSearchItem'
        page:
          type: integer
          format: int64
        pages:
          type: integer
          format: int64
        per_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    AlertStatusResponse:
      type: object
      description: |-
        Documentation-only mirror of `update_alert_status`'s `serde_json::json!`
        body.
      required:
      - message
      - alert_id
      - new_status
      properties:
        alert_id:
          type: integer
          format: int32
        message:
          type: string
        new_status:
          type: string
    AlertUpdateResponse:
      type: object
      description: Documentation-only mirror of `update_alert`'s `serde_json::json!` body.
      required:
      - message
      - alert
      properties:
        alert:
          $ref: '#/components/schemas/AlertUpdateSummary'
        message:
          type: string
    AlertUpdateSummary:
      type: object
      description: Summary embedded in [`AlertUpdateResponse`].
      required:
      - id
      - title
      - severity
      - status
      properties:
        id:
          type: integer
          format: int32
        severity:
          type: string
        status:
          type: string
        title:
          type: string
        updated_at:
          type:
          - string
          - 'null'
...
    ErrorResponse:
      type: object
      description: |-
        Documentation-only mirror of `ApiError`'s bare `{"error": msg}` wire shape
        (`BadRequest`/`Unauthorized`/`Forbidden`/`NotFound`/`Conflict` in this
        service all use it, when their body happens to be bare) — `ApiError`
        itself builds its body with `serde_json::json!` rather than a typed
        struct, so this type exists solely to give `utoipa` something to
        reference in `#[utoipa::path]` `responses(...)` clauses. See
        `docs/v2-port/openapi-pattern.md`.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    RefreshRequest:
      type: object
      description: POST /auth/refresh body.
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      description: Documentation-only mirror of `refresh`'s `serde_json::json!` body.
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          format: int64
          description: Access-token lifetime in seconds.
        refresh_token:
          type: string
        token_type:
          type: string
...
    SearchBody:
      type: object
      description: |-
        IOCSearchRequest — `tags` is accepted for schema parity but v1 never
        filters on it, and neither do we.
      properties:
        confidence_min:
          type:
          - number
          - 'null'
          format: double
        include_expired:
          type:
          - boolean
          - 'null'
        indicator_type:
          type:
          - array
          - 'null'
          items:
            type: string
        page:
          type:
          - integer
          - 'null'
          format: int64
        per_page:
          type:
          - integer
          - 'null'
          format: int64
        query:
          type:
          - string
          - 'null'
        source:
          type:
          - string
          - 'null'
        tags:
          type:
          - array
          - 'null'
          items:
            type: string
        threat_level:
          type:
          - array
          - 'null'
          items:
            type: string
...
    StatusBody:
      type: object
      description: PUT /alerts/{alert_id}/status body.
      properties:
        status:
          type:
          - string
          - 'null'
...
    UpdateBody:
      type: object
      description: AlertUpdateRequest — every field optional; absent fields are not touched.
      properties:
        assigned_to:
          type:
          - integer
          - 'null'
          format: int32
        description:
          type:
          - string
          - 'null'
        resolution_notes:
          type:
          - string
          - 'null'
        severity:
          type:
          - string
          - 'null'
        status:
          type:
          - string
          - 'null'
        title:
          type:
          - string
          - 'null'
...
    ValidationErrorResponse:
      type: object
      description: |-
        Documentation-only mirror of `ApiError::Validation`'s wire shape — a
        bare error message plus a list of per-field validation failure objects
        (`{loc, msg, type}`, matching the `ApiJson` rejection shape below).
      required:
      - error
      - details
      properties:
        details:
          type: array
          items: {}
          description: 'Per-field failure details, each `{loc: [...], msg: ..., type: ...}`.'
        error:
          type: string
          description: Always `"Validation error"`.
  securitySchemes:
    bearer_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    endpoint_hmac:
      type: apiKey
      in: header
      name: X-API-Key
...